+++ src/com/onelogin/fractions/solver/SubtractionSolver.java
@@ -8,16 +8,16 @@ public class SubtractionSolver extends Solver {
     Fraction simpleFirstTerm = convertMixedToSimpleFraction(firstTerm);
     Fraction simpleSecondTerm = convertMixedToSimpleFraction(secondTerm);
 
-    int numerator = 1;
-    int denominator = 1;
+    long numerator = 1;
+    long denominator = 1;
 
     if (simpleFirstTerm.getDenominator() == simpleSecondTerm.getDenominator()) {
       denominator = simpleFirstTerm.getDenominator();
       numerator = simpleFirstTerm.getNumerator() - simpleSecondTerm.getNumerator();
     } else {
       denominator = simpleFirstTerm.getDenominator() * simpleSecondTerm.getDenominator();
-      int firstMult = simpleFirstTerm.getNumerator() * simpleSecondTerm.getDenominator();
-      int secondMult = simpleFirstTerm.getDenominator() * simpleSecondTerm.getNumerator();
+      long firstMult = simpleFirstTerm.getNumerator() * simpleSecondTerm.getDenominator();
+      long secondMult = simpleFirstTerm.getDenominator() * simpleSecondTerm.getNumerator();
       numerator = firstMult - secondMult;
     }
